refactor: cleave location computation from sampleView
